+++ src/sys/device/display/x11/error.rs
@@ -7,7 +7,11 @@ use crate::{Display, Error, FmtResult, Formatter};
 
 #[doc = crate::_tags!(unix error)]
 /// XCB/X11 error categories.
-#[doc = crate::_doc_meta!{location("sys/device/display/x11")}]
+#[doc = crate::_doc_meta!{
+    location("sys/device/display/x11"),
+    #[cfg(target_pointer_width = "64")]
+    test_size_of(XError = 24|192; niche Option),
+}]
 ///
 /// Represents the main classes of failures that can occur when interacting with
 /// the X server through XCB. These variants cover connection issues, setup failures,