+++ src/sys/device/display/x11/event.rs
@@ -10,7 +10,11 @@ use crate::{EventButton, EventButtonState, EventButtons, EventKey, EventTimestam
 
 #[doc = crate::_tags!(unix event)]
 /// Wrapper for an XCB event.
-#[doc = crate::_doc_meta!{location("sys/device/display/x11")}]
+#[doc = crate::_doc_meta!{
+    location("sys/device/display/x11"),
+    #[cfg(target_pointer_width = "64")]
+    test_size_of(XEvent = 8|64),
+}]
 #[derive(Debug)]
 pub struct XEvent {
     pub(super) raw: *mut _raw::xcb_generic_event_t,