+++ src/sys/device/display/x11/surface/cpu.rs
@@ -7,7 +7,11 @@ use crate::{Vec, XDisplay, XError, XImageStore, vec_ as vec};
 
 #[doc = crate::_tags!(unix runtime)]
 /// Cpu-backed pixel buffer for X11 image upload.
-#[doc = crate::_doc_meta!{location("sys/device/display/x11")}]
+#[doc = crate::_doc_meta!{
+    location("sys/device/display/x11"),
+    #[cfg(target_pointer_width = "64")]
+    test_size_of(XCpuBuffer = 32|256; niche Option),
+}]
 ///
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
 pub struct XCpuBuffer {